% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_archive.R
\name{as_tibble.archive}
\alias{as_tibble.archive}
\title{Convert an archive manifest to a tibble}
\usage{
\method{as_tibble}{archive}(x, ...)
}
\arguments{
\item{x}{An \code{archive}.}

\item{...}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}
}
\value{
The archive manifest as a tibble.
}
\description{
Convert an archive manifest to a tibble
}
