% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_run.R
\name{as_tibble.audit}
\alias{as_tibble.audit}
\title{Coerce an audit to a tibble}
\usage{
\method{as_tibble}{audit}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=audit_run]{audit_run()}} result.}

\item{...}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}
}
\value{
The per-cell tibble.
}
\description{
Coerce an audit to a tibble
}
