% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder_run.R
\name{as_tibble.coded_corpus}
\alias{as_tibble.coded_corpus}
\title{Coerce a coded corpus to a tibble}
\usage{
\method{as_tibble}{coded_corpus}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=code_corpus]{code_corpus()}} result.}

\item{...}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}
}
\value{
The coded corpus's \code{data} tibble.
}
\description{
Coerce a coded corpus to a tibble
}
