% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder_codebook.R
\name{codebook_labels}
\alias{codebook_labels}
\title{Labels defined by a codebook}
\usage{
codebook_labels(x)
}
\arguments{
\item{x}{A \code{\link[=codebook]{codebook()}}.}
}
\value{
Character vector of category labels.
}
\description{
Labels defined by a codebook
}
