% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_persist.R
\name{archive_read}
\alias{archive_read}
\title{Read an archive from a directory}
\usage{
archive_read(dir)
}
\arguments{
\item{dir}{A directory written by \code{\link[=archive_write]{archive_write()}}.}
}
\value{
An \code{archive}. Run \code{\link[=archive_check]{archive_check()}} right after: reading and
verifying are deliberately separate acts.
}
\description{
Read an archive from a directory
}
